+++ utils/dynstream/event_queue.go
@@ -82,7 +82,7 @@ func (q *eventQueue[A, P, T, D, H]) releasePath(path *pathInfo[A, P, T, D, H]) {
 	}
 
 	if path.areaMemStat != nil {
-		path.areaMemStat.decPendingSize(path, int64(path.pendingSize.Load()))
+		path.areaMemStat.decPendingSize(path, path.pendingSize.Load())
 		path.areaMemStat.lastSizeDecreaseTime.Store(time.Now())
 	}
